Move to use wix site instead of firestore
firestore is limiting us in the number of calls we can make to it per day (50K) in its free plan.
I initially thought of bypassing it by making the next.js app on vercel cache the data on load, so only a single call is being made, but vercels calls are serverless, so cache cannot bbe used between requests.

Wix data provides much better service, where its free plan is only limited to 500MB (way above our needs) and does not limit the number of calls.

I added wix code files here so we can keep track of changes and also to make it truely open-source, so other can make use of this, if they want.