Key results by index
Results only change when data is loading, but some pages might have the same title (due to migration), so we better key them by index.

There seems to be another problem caused when some results contain images from the old site with relative URL, which causes next.js to return an 404 error that hides our `body`. But this only happens in dev and not in production, so we can ignore.