Fix quotes handling
In commit db4f0bfe05bffe680a95a9cb789f3918dc62e61c we escaped " to %22 when cpying a URL in order for whatsapp to recognize it properly as URL, but this causes the URL to not work when clicking it, as %22 was escaped to %2522.
To fix it, we now replace " with \u201d (Right Double Quotation Mark), which looks almost the same, but makes whatsapp think it is a valid URL character, and when opening the UR, we replace it back to regular double-quotes (").